requirements.txt
pymupdf==1.26.7
anthropic==0.76.0
python-dotenv==1.2.1
